Might handle XML-per-file

+++ clean_studio_xml.py
@@ -12,6 +12,12 @@ import sys
 import xml.etree
 import xml.etree.ElementTree as ET
 
+
+parser = argparse.ArgumentParser(description = "Clean up XML spat out by Studio.")
+parser.add_argument("base", help="Base directory of Studio-dumped XML")
+args = parser.parse_args()
+
+
 shre = re.compile("^[a-f0-9]+$")
 def studio_hash(s):
     ''' Check if a string is a Studio-generated hash '''
@@ -58,17 +64,30 @@ def unique_url_slug(s):
     used_url_names.add(new_string)
     return new_string
 
+display_map = {}
+def set_url_name_slug(e, new_name):
+    ''' Maintain a mapping from new URL names to old ones for e.g. analytics. 
+    We'll keep the mapping in static files. 
+    '''
+    new_name = unique_url_slug(new_name)
+    display_map[new_name] = e.attrib['url_name']
+    e.attrib['url_name'] = new_name
 
-parser = argparse.ArgumentParser(description = "Clean up XML spat out by Studio.")
-parser.add_argument("base", help="Base directory of Studio-dumped XML")
-args = parser.parse_args()
-
-# given element of the form <foo url_name="...">, if there's a directory named "tag",
-# parse the file named by the "url_name" attribute and add subtree as a child of element.
-# Recursively load each subtree, add parent pointers so we can walk up the tree.
 def load_subtree(element):
+    ''' given element of the form <foo url_name="...">, if there's a directory named "tag",
+    parse the file named by the "url_name" attribute and add subtree as a child of element.
+    Recursively load each subtree, add parent pointers so we can walk up the tree. '''
+    if 'url_name' not in element.attrib:
+        return
+
+    filename = os.path.join(args.base, element.tag,element.attrib['url_name']+'.xml')
+
     if os.path.isdir(os.path.join(args.base, element.tag)) and element.attrib.has_key('url_name'):
-        filename = os.path.join(args.base, element.tag,element.attrib['url_name']+'.xml')
+        if not os.path.exists(filename):
+            for child in element: 
+                child.parent = element
+                load_subtree(child)
+            return
         subtree = ET.parse(os.path.join(args.base, element.tag,element.attrib['url_name']+'.xml')).getroot()
         os.unlink(filename)
         for child in subtree:
@@ -93,19 +112,10 @@ def load_subtree(element):
             element.append(subtree)
             subtree.parent = element
 
-# starting with element, move up the DOM tree looking for tag
-def find_parent(element,tag):
-    while True:
-        if element is None or element.tag == tag:
-            return element
-        element = element.parent
-
-# get root of course XML tree
+# get root of course XML tree and load the XML for the entire course
 tree = ET.parse(os.path.join(args.base, 'course.xml'))
 root = tree.getroot()
 root.parent = None
-
-# load the XML for the entire course
 load_subtree(root)
 
 ## Now, we'll clean up the URL names Studio assigned
@@ -118,7 +128,7 @@ for e in tree.iter():
 # change the URL name to be a sluggification of the display name
 for e in tree.iter():
     if 'display_name' in e.attrib and studio_hash(e.attrib['url_name']):
-        e.attrib['url_name'] = unique_url_slug(e.attrib['display_name'])
+        set_url_name_slug(e, e.attrib['display_name'])
 
 ## Step 3: We'll clean up the filenames Studio assigned for our HTML files
 for e in tree.iter():
@@ -141,15 +151,25 @@ for e in tree.iter():
         if related_index >= 0: 
             related_node = e.parent._children[related_index]
             if studio_hash(e.attrib['url_name']):
-                e.attrib['url_name'] = related_node.attrib['url_name'] + '_discussion'
+                set_url_name_slug(e, related_node.attrib['url_name'] + '_discussion')
             if not 'discussion_target' in e.attrib or studio_hash(e.attrib['discussion_target']):
                 if 'display_name' in related_node.attrib:
                     e.attrib['discussion_target'] = related_node.attrib['display_name']
 
-# Step 5: We'll dump back to the file system
-output = ET.tostring(root) # TODO: Tounicode
 
-#output_file = open(args.output, "w")
+# We're done. Dump course.xml back to the file system
+output = ET.tostring(root) # TODO: Tounicode
 output_file = open(os.path.join(args.base, 'course.xml'), "w")
 output_file.write(output)
 output_file.close()
+
+# And finally, dump the mapping file
+mapping_filename = os.path.join(args.base, 'static/urlname_mapping.json')
+i = 0
+while os.path.exists(mapping_filename):
+    mapping_filename = os.path.join(args.base, 'static/urlname_mapping_{i}.json'.format(i=i))
+    i = i+1
+
+mapping_file = open(mapping_filename, "w")
+mapping_file.write(json.dumps(display_map, indent=2))
+mapping_file.close()
